Make the developer code actually runnable: 1) Moved everything under src/main/java... as things were spread throughout multiple roots 2) Cleaned up some files 3) There were two Robot.java in two different roots, so I deleted the one that didn't work with the current code.
