+++ src/main/java/com/team2502/robot2017/command/DriveStraightCommand.java
@@ -1,11 +1,8 @@
 package com.team2502.robot2017.command;
 
 import com.team2502.robot2017.Robot;
-import com.team2502.robot2017.RobotMap;
 import com.team2502.robot2017.subsystem.DriveTrainSubsystem;
-
 import com.team2502.robot2017.subsystem.ShooterSubsystem;
-
 import edu.wpi.first.wpilibj.Encoder;
 import edu.wpi.first.wpilibj.PIDController;
 import edu.wpi.first.wpilibj.PIDOutput;
