+++ src/main/java/com/team2502/robot2017/command/DriveTillDistanceCommand.java
@@ -3,7 +3,6 @@ package com.team2502.robot2017.command;
 import com.team2502.robot2017.Robot;
 import com.team2502.robot2017.subsystem.DistanceSensorSubsystem;
 import com.team2502.robot2017.subsystem.DriveTrainSubsystem;
-
 import edu.wpi.first.wpilibj.command.Command;
 
 public class DriveTillDistanceCommand extends Command
