+++ src/main/java/com/team2502/robot2017/command/GearCommandPushGear.java
@@ -1,8 +1,8 @@
 package com.team2502.robot2017.command;
 
 import com.team2502.robot2017.Robot;
-import edu.wpi.first.wpilibj.command.Command;
 import com.team2502.robot2017.subsystem.GearBoxSubsystem;
+import edu.wpi.first.wpilibj.command.Command;
 
 public class GearCommandPushGear extends Command
 {
