+++ src/main/java/com/team2502/robot2017/command/NavXMoveCommand.java
@@ -4,7 +4,6 @@ import com.kauailabs.navx.frc.AHRS;
 import com.team2502.robot2017.Robot;
 import com.team2502.robot2017.subsystem.DistanceSensorSubsystem;
 import com.team2502.robot2017.subsystem.DriveTrainSubsystem;
-
 import edu.wpi.first.wpilibj.command.Command;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
