+++ src/main/java/com/team2502/robot2017/command/autonomous/EncDriveDistanceCommand.java
@@ -2,7 +2,6 @@ package com.team2502.robot2017.command.autonomous;
 
 import com.team2502.robot2017.Robot;
 import com.team2502.robot2017.subsystem.DriveTrainSubsystem;
-import com.team2502.robot2017.subsystem.GearBoxSubsystem;
 import edu.wpi.first.wpilibj.command.Command;
 
 public class EncDriveDistanceCommand extends Command
