+++ src/main/java/com/team2502/robot2017/command/autonomous/FeedShooterCommand.java
@@ -1,7 +1,6 @@
 package com.team2502.robot2017.command.autonomous;
 
 import com.team2502.robot2017.Robot;
-import com.team2502.robot2017.subsystem.GearBoxSubsystem;
 import com.team2502.robot2017.subsystem.ShooterSubsystem;
 import edu.wpi.first.wpilibj.command.Command;
